+++ intelmq/lib/upgrades.py
@@ -802,6 +802,10 @@ def v310_feed_changes(configuration, harmonization, dry_run, **kwargs):
     found_abusech_removed_parsers = []
     found_abusech_feodotracker_csv = []
     found_abusech_feodotracker_browse = []
+    found_viriback = []
+    found_netlab_mirai_scanner = []
+    found_benkow_panels = []
+    found_taichung = []
     messages = []
     for bot_id, bot in configuration.items():
         if bot_id == 'global':
@@ -818,6 +822,14 @@ def v310_feed_changes(configuration, harmonization, dry_run, **kwargs):
                 found_abusech_feodotracker_browse.append(bot_id)
             if http_url.startswith("https://feodotracker.abuse.ch/downloads/ipblocklist.csv"):
                 found_abusech_feodotracker_csv.append(bot_id)
+            if http_url == "http://tracker.viriback.com/":
+                found_viriback.append(bot_id)
+            if http_url.startswith("http://data.netlab.360.com/feeds/mirai-scanner/scanner.list"):
+                found_netlab_mirai_scanner.append(bot_id)
+            if "benkow.cc/export.php" in http_url:  # both HTTP and HTTPS
+                found_benkow_panels.append(bot_id)
+            if http_url.startswith("https://www.tc.edu.tw/net/netflow/lkout/recent"):
+                found_taichung.append(bot_id)
         if bot["module"] == "intelmq.bots.parsers.autoshun.parser":
             found_autoshun.append(bot_id)
         if bot["module"] == "intelmq.bots.parsers.dshield.parser_domain":
@@ -830,6 +842,8 @@ def v310_feed_changes(configuration, harmonization, dry_run, **kwargs):
                 "classification.type": bot["parameters"]["type"]
             }
             del bot["parameters"]["type"]
+        if bot["module"] == "intelmq.bots.parsers.taichung.parser":
+            found_taichung.append(bot_id)
 
     if found_malc0de:
         messages.append('A discontinued feed "Malc0de" has been found '
@@ -854,6 +868,24 @@ def v310_feed_changes(configuration, harmonization, dry_run, **kwargs):
     if found_abusech_removed_parsers:
         messages.append('A discontinued bot module has been found'
                         'as bot %s.' % ', '.join(sorted(found_abusech_removed_parsers)))
+
+    if found_viriback:
+        messages.append('The feed "Viriback Unsafe Site" has been replaced. Please see the feed'
+                        ' "Viriback C2 Tracker" and'
+                        ' adjust your configuration. Affected bots: %s' % ', '.join(found_viriback))
+
+    if found_netlab_mirai_scanner:
+        messages.append('A discontinued feed "Netlab Mirai Scanner" has been found '
+                        'as bot %s.' % ', '.join(sorted(found_netlab_mirai_scanner)))
+
+    if found_benkow_panels:
+        messages.append('The feed "Benkow Malware Panels Tracker" has been changed. Please see the feed\'s'
+                        ' documentation and adjust your configuration.'
+                        ' Affected bots: %s' % ', '.join(found_benkow_panels))
+
+    if found_taichung:
+        messages.append('A discontinued feed "Taichung" has been found '
+                        'as bot %s.' % ', '.join(sorted(found_taichung)))
 
     messages = ' '.join(messages)
     return messages + ' Remove affected bots yourself.' if messages else None, configuration, harmonization