+++ CHANGELOG.md
@@ -143,6 +143,7 @@ CHANGELOG
   - Adds fields from the JSON feed.
 - `intelmq.bots.parsers.generic.parser_csv`: Parameter `type` is deprecated, `default_fields` should be used. (PR#2293 by Filip Pokorný)
 - `intelmq.bots.parsers.generic.parser_csv`: Parameter `skip_header` now allows also integer as a fixed number of lines to skip. (PR#2313 by Filip Pokorný)
+- `intelmq.bots.parsers.taichung.parser`: Removed (PR#2266 by Filip Pokorný)
 
 #### Experts
 - `intelmq.bots.experts.domain_valid`: New bot for checking domain's validity (PR#1966 by Marius Karotkis).
@@ -170,6 +171,10 @@ CHANGELOG
 - Fix wrong operator for list-contains-value operation in sieve expert documentation (PR#2256 by Filip Pokorný).
 - Added documentation on `default_fields` parameter (PR#2293 by Filip Pokorný).
 - Updated documentation on `skip_header` parameter (PR#2313 by Filip Pokorný).
+- Viriback Unsafe Sites feed replaced with Viriback C2 Tracker. (PR#2266 by Filip Pokorný)
+- Netlab 360 Mirai Scanner feed removed as it is discontinued. (PR#2266 by Filip Pokorný)
+- Benkow Malware Panels Tracker feed changed parser configuration. (PR#2266 by Filip Pokorný)
+- Taichung feed removed as it is discontinued. (PR#2266 by Filip Pokorný)
 
 ### Packaging
 - Remove deleted `intelmq.bots.experts.sieve.validator` from executables in `setup.py` (PR#2256 by Filip Pokorný).